Refactor Buff constructor to take full config object
The constructor for the Buff class has been revised to take in the full config object instead of individual parameters. The constructor's parameters (request_interval and request_kwargs) have been replaced with a single config parameter. This helps to simplify code and improve the maintainability. Also, a new attribute 'sort_by' is added to the Buff class and the relevant configuration is provided(#28).

Signed-off-by: hldh214 <hldh214@gmail.com>