Add retry mechanism to Buff request method(#28)
Added a retry mechanism to the Buff class's request method, in the event of a BuffError exception, using the tenacity library. It is set to try a maximum of 5 times, utilizing a random exponential wait strategy with a minimum of 8 seconds and maximum of 60 seconds between attempts, enhancing error tolerance and stability of data fetching operations.

Signed-off-by: hldh214 <hldh214@gmail.com>